12.21 commit because i will sleep,finsh more

+++ src/main/java/com/hsyuan/inventropy/entity/Project.java
@@ -14,7 +14,7 @@ public class Project {
     private String projectName;
     private String projectType;
     private String applicant;
-    private Long applicantId;
+    private Integer applicantId;
     private Double funds;
     private Double remainingFunds;
     private String content;
